refactor combobox into model picker. it was not generic
