--- src/lib/components/ui/combobox/index.ts
@@ -1,3 +1,0 @@
-import Root from "./combobox.svelte";
-
-export { Root as Combobox };
